[DEV] don't publish wahooMapsCreator as .zip and don't mention in docs
not the whole repo is needed to download, only the gdal-user.yml would be needed. wahoomc is installed via the .yml file. If one wants to change coding, the developer-env should be used

BREAKING CHANGE:
Users who downloaded the .zip files from the release page should migrate to the PyPI version and update the python module accordingly.