Minimalistic maps selection and other zoom interval
Minimalistic map selection uses smaller maps when available resulting in potentially smaller downloads and faster processing (it is faster to split tiles from smaller files)

And changed to zoom-interval-conf from 10,0,17 to 12,0,17. This creates smaller sub-files in the resulting maps. This in turn makes sure that for instance Paris can be rendered on the Bolt 1&2 and the roam 1&2