[FEAT] Minimalistic raw maps: Download regions instead of countries (#275)
* Minimalistic maps selection and other zoom interval

Minimalistic map selection uses smaller maps when available resulting in potentially smaller downloads and faster processing (it is faster to split tiles from smaller files)

And changed to zoom-interval-conf from 10,0,17 to 12,0,17. This creates smaller sub-files in the resulting maps. This in turn makes sure that for instance Paris can be rendered on the Bolt 1&2 and the roam 1&2

* Removed extra spaces

* refactor, delete comments

supported by Copilot ;-)

* undo changing zoom-interval-conf

- the change is separated from this branch/PR and exists now here: https://github.com/treee111/wahooMapsCreator/commit/792f6c7acd2d9b1aefcc28e1c91150a8bcd43a99

* update unittests by regions instead of countries

- due to processing smaller parts/ regions instead of whole countries the border-countries/regions, involved countries/regions change

* update comment to reflect actual coding

---------

Co-authored-by: Benjamin Kreuscher <benni.kreuscher@gmail.com>