[FEAT] Add GUI & CLI switch for mapwriter hdd mode (#280)
* Add GUI & CLI switch for mapwriter hdd mode

Added a switch in the advance GUI settings and the CLI to enable the mapsforge mapwriter hdd mode. If you run out of memory during map creation you can try this option. As indicated it tries to use the hdd  as memory instead of using ram.
The obvious downside to this is it' slower. In my test with Malta using ram mode took 91 seconds in hdd mode it took 158 seconds

* be consistent with the name of the argument in cli arguments and o_input_data

* add unittest for mapwriter hdd mode

---------

Co-authored-by: Benjamin Kreuscher <benni.kreuscher@gmail.com>